move scroll to overlay
