+++ src/lib/core/core.ts
@@ -6,8 +6,6 @@ import {PortalModule} from './portal/portal-directives';
 import {OverlayModule} from './overlay/overlay-directives';
 import {A11yModule, A11Y_PROVIDERS} from './a11y/index';
 import {OVERLAY_PROVIDERS} from './overlay/overlay';
-import {ScrollDispatcher} from './scroll/scroll-dispatcher';
-import {ScrollModule} from './scroll/scrollable';
 
 
 // RTL
@@ -49,8 +47,7 @@ export {
 } from './overlay/overlay-directives';
 export * from './overlay/position/connected-position-strategy';
 export * from './overlay/position/connected-position';
-export * from './scroll/scrollable';
-export * from './scroll/scroll-dispatcher';
+export {ScrollDispatcher} from './overlay/scroll/scroll-dispatcher';
 
 // Gestures
 export {GestureConfig} from './gestures/gesture-config';
@@ -111,33 +108,30 @@ export {coerceNumberProperty} from './coercion/number-property';
 export {DefaultStyleCompatibilityModeModule} from './compatibility/default-mode';
 export {NoConflictStyleCompatibilityMode} from './compatibility/no-conflict-mode';
 
-// ScrollDispatcher
-export {ScrollDispatcher} from './scroll/scroll-dispatcher';
-export {Scrollable} from './scroll/scrollable';
 
 @NgModule({
-  imports: [MdLineModule,
+  imports: [
+    MdLineModule,
     RtlModule,
     MdRippleModule,
     PortalModule,
     OverlayModule,
     A11yModule,
-    ScrollModule
   ],
-  exports: [MdLineModule,
+  exports: [
+    MdLineModule,
     RtlModule,
     MdRippleModule,
     PortalModule,
     OverlayModule,
     A11yModule,
-    ScrollModule
   ],
 })
 export class MdCoreModule {
   static forRoot(): ModuleWithProviders {
     return {
       ngModule: MdCoreModule,
-      providers: [A11Y_PROVIDERS, OVERLAY_PROVIDERS, ScrollDispatcher],
+      providers: [A11Y_PROVIDERS, OVERLAY_PROVIDERS],
     };
   }
 }
